+++ lib/src/mapper/mapper.dart
@@ -18,7 +18,13 @@ class ConferenceMapper {
     var status =
         ConferenceStatus.decode(map["status"] as String? ?? "DEFAULT") ??
             ConferenceStatus.defaultStatus;
-    return Conference(alias, id, isNew, participants, status);
+    var spatialAudioStyle =
+        map.containsKey("spatialAudioStyle") && map["spatialAudioStyle"] != null
+            ? SpatialAudioStyle.decode(map["spatialAudioStyle"] as String)
+            : null;
+
+    return Conference(
+        alias, id, isNew, participants, status, spatialAudioStyle);
   }
 
   static List<Participant> prepareParticipantsList(List<Object?> participants) {