+++ lib/src/sdk_api/command_service.dart
@@ -6,6 +6,8 @@ import 'models/enums.dart';
 import 'models/events.dart';
 
 /// The CommandService allows the application to send and receive text messages and notifications during a conference.
+///
+/// {@category Services}
 class CommandService {
   /// @internal
   final _methodChannel =