+++ lib/src/sdk_api/conference_service.dart
@@ -10,6 +10,8 @@ import 'models/spatial.dart';
 import 'session_service.dart';
 
 /// The ConferenceService allows an application to manage a conference life-cycle and interact with the conference. The service allows creating, joining, and leaving conferences and managing the audio, video, and screen-share streams.
+///
+/// {@category Services}
 class ConferenceService {
   /// @internal
   final _methodChannel =
@@ -275,36 +277,41 @@ class ConferenceService {
   /// This method uses the following parameters:
   /// - [max]: The maximum number of video streams that may be transmitted to the local participant. The valid parameter values are between 0 and 4. By default, the parameter is set to 4.
   /// - [prioritizedParticipants]: The list of the prioritized participants. This parameter allows using a pin option to prioritize specific participant's video streams and display their videos even when these participants do not talk.
-  Future<bool?> setMaxVideoForwarding(
+  @Deprecated(
+      'This method is supported only in SDK 3.6.0. To set Video Forwarding in SDK 3.6.1 and later, use the [ConferenceService.setVideoForwarding()] method.')
+  Future<bool> setMaxVideoForwarding(
       MaxVideoForwarding max, List<Participant> prioritizedParticipants) async {
-    return await _methodChannel.invokeMethod<bool>("setMaxVideoForwarding", {
+    final result =
+        await _methodChannel.invokeMethod<bool>("setMaxVideoForwarding", {
       "max": max,
       "prioritizedParticipants":
           prioritizedParticipants.map((e) => e.toJson()).toList()
     });
+    return Future.value(result);
   }
 
   ///Sets the maximum number of video streams that may be transmitted to the local participant.
   ///For more information, see the
   ///[Video Forwarding](https://docs.dolby.io/communications-apis/docs/guides-video-forwarding) article.
   ///
-  ///[param] strategy Defines how the SDK should select conference participants whose videos will be
+  /// This method uses the following parameters:
+  /// - [strategy]: Defines how the SDK should select conference participants whose videos will be
   ///transmitted to the local participant. There are two possible values; the selection can be either
   ///based on the participants' audio volume or the distance from the local participant
-  ///[max] The maximum number of video streams that may be transmitted to the local
+  /// - [max]: The maximum number of video streams that may be transmitted to the local
   ///participant. The valid values are between 0 and 4. The default value is 4.
   ///In the case of providing a value smaller than 0 or greater than 4, SDK triggers
-  ///the [IllegalStateException] error.
-  ///[prioritizedParticipants] The list of participants' objects. Allows prioritizing specific participant's
+  ///an error.
+  /// - [prioritizedParticipants]: The list of participants' objects. Allows prioritizing specific participant's
   ///video streams and display their videos even when these participants do not talk.
   ///For example, in the case of virtual classes, this option allows participants to pin the teacher's
   ///video and see the teacher, even when the teacher is not the active speaker.
-  Future<bool?> setVideoForwarding(
+  Future<bool> setVideoForwarding(
     VideoForwardingStrategy strategy,
     MaxVideoForwarding max,
     List<Participant> prioritizedParticipants,
   ) async {
-    return await _methodChannel.invokeMethod<bool>(
+    final result = await _methodChannel.invokeMethod<bool>(
       "setVideoForwarding",
       {
         "strategy": strategy.encode(),
@@ -313,6 +320,7 @@ class ConferenceService {
             prioritizedParticipants.map((e) => e.toJson()).toList()
       },
     );
+    return Future.value(result ?? false);
   }
 
   /// Gets the Participant object based on the [participantId].