+++ lib/src/sdk_api/media_device_service.dart
@@ -2,6 +2,8 @@ import '../dolbyio_comms_sdk_flutter_platform_interface.dart';
 import 'models/enums.dart';
 
 /// The MediaDeviceService allows an application to manage media devices that are used during a conference.
+///
+/// {@category Services}
 class MediaDeviceService {
   /// @internal
   final _methodChannel = DolbyioCommsSdkFlutterPlatform.createMethodChannel(
@@ -19,7 +21,7 @@ class MediaDeviceService {
         await _methodChannel.invokeMethod<bool>("isFrontCamera"));
   }
 
-  /// Sets the [comfort noise level] for output devices in Dolby Voice conferences.
+  /// Sets the comfort noise level ([noiseLevel]) for output devices in Dolby Voice conferences.
   Future<void> setComfortNoiseLevel(ComfortNoiseLevel noiseLevel) async {
     return await _methodChannel.invokeMethod<void>(
         "setComfortNoiseLevel", {"noiseLevel": noiseLevel.encode()});
@@ -30,7 +32,7 @@ class MediaDeviceService {
     return await _methodChannel.invokeMethod<void>("switchCamera");
   }
 
-  /// Switches the current speaker to a different speaker that is available. The method is available only on iOS. Using the method on Android triggers an error.
+  /// Switches the current speaker to a different speaker that is available.
   Future<void> switchSpeaker() async {
     return await _methodChannel.invokeMethod<void>("switchSpeaker");
   }