+++ lib/src/sdk_api/models/enums.dart
@@ -3,6 +3,8 @@ abstract class EnumWithStringValue implements Enum {
 }
 
 /// The ComfortNoiseLevel enum gathers the available comfort noise levels.
+///
+/// {@category Models}
 enum ComfortNoiseLevel {
   /// The default comfort noise level that is based on the device database. The database contains the proper comfort noise levels, individual for all devices.
   defaultLevel('default'),
@@ -34,6 +36,8 @@ enum ComfortNoiseLevel {
 }
 
 /// The FilePresentationServiceEventNames enum gathers events informing about the file presentation status.
+///
+/// {@category Models}
 enum FilePresentationServiceEventNames implements EnumWithStringValue {
   /// Emitted when a file is converted.
   fileConverted('EVENT_FILEPRESENTATION_FILE_CONVERTED'),
@@ -65,6 +69,8 @@ enum FilePresentationServiceEventNames implements EnumWithStringValue {
 }
 
 /// The NotificationServiceEventNames enum gathers the NotificationService events.
+///
+/// {@category Models}
 enum NotificationServiceEventNames implements EnumWithStringValue {
   /// Emitted when an application user receives an invitation.
   invitationReceived('EVENT_NOTIFICATION_INVITATION_RECEIVED');
@@ -87,6 +93,8 @@ enum NotificationServiceEventNames implements EnumWithStringValue {
 }
 
 /// The CommandServiceEventNames enum gathers the CommandService events.
+///
+/// {@category Models}
 enum CommandServiceEventNames implements EnumWithStringValue {
   /// Emitted when a participant receives a message.
   messageReceived('EVENT_COMMAND_MESSAGE_RECEIVED');
@@ -109,6 +117,8 @@ enum CommandServiceEventNames implements EnumWithStringValue {
 }
 
 /// The VideoPresentationState enum gathers the possible statuses of a video presentation.
+///
+/// {@category Models}
 enum VideoPresentationState {
   /// The video presentation is paused.
   paused('paused'),
@@ -133,6 +143,8 @@ enum VideoPresentationState {
 }
 
 /// The VideoPresentationEventNames enum gathers the possible statuses of a video presentation.
+///
+/// {@category Models}
 enum VideoPresentationEventNames implements EnumWithStringValue {
   /// Emitted when a video presentation is paused.
   videoPresentationPaused('EVENT_VIDEOPRESENTATION_PAUSED'),