+++ lib/src/sdk_api/models/file_presentation.dart
@@ -1,6 +1,8 @@
 import 'participant.dart';
 
 /// The File class gathers information about a file that a presenter wants to share during a conference.
+///
+/// {@category Models}
 class File {
   /// The local path of a file.
   String uri;
@@ -13,6 +15,8 @@ class File {
 }
 
 /// The FileConverted class gathers information about a converted file.
+///
+/// {@category Models}
 class FileConverted {
   /// The file ID.
   String id;
@@ -41,6 +45,8 @@ class FileConverted {
 }
 
 /// The FilePresentation class gathers information about a file presentation.
+///
+/// {@category Models}
 class FilePresentation {
   /// The file ID.
   String id;