+++ lib/src/sdk_api/models/recording.dart
@@ -2,6 +2,8 @@
 import 'package:collection/collection.dart';
 
 /// The RecordingInformation class gathers information about a conference recording.
+///
+/// {@category Models}
 class RecordingInformation {
   /// The ID of the participant who started the recording.
   String? participantId;
@@ -20,6 +22,8 @@ class RecordingInformation {
 }
 
 /// The RecordingStatus enum gathers the possible statuses of recording.
+///
+/// {@category Models}
 enum RecordingStatus {
   /// The conference is recorded.
   recording('RECORDING'),