+++ lib/src/sdk_api/models/spatial.dart
@@ -7,6 +7,8 @@
 /// - Z-axis: Extends positive forwards
 ///
 /// The [ConferenceService.setSpatialEnvironment] method allows the application to choose the meaning of each axis and match the usage of the application.
+///
+/// {@category Models}
 class SpatialPosition {
   /// The x-coordinate of a new audio location.
   double x;
@@ -26,6 +28,8 @@ class SpatialPosition {
 /// The SpatialScale class defines how to convert units from the application's coordinate system (pixels or centimeters) into meters used by the spatial audio coordinate system. For example, let's assume that SpatialScale is set to (100,100,100), which indicates that 100 of the applications units (cm) map to 1 meter for the audio coordinates. If the listener's location is (0,0,0)cm and a remote participant's location is (200,200,200)cm, the listener has an impression of hearing the remote participant from the (2,2,2)m location.
 ///
 ///**Note**: A scale value must have a value greater than zero. The default is (1,1,1).
+///
+/// {@category Models}
 class SpatialScale {
   /// The x component of the SpatialScale vector.
   double x;
@@ -53,6 +57,8 @@ class SpatialScale {
 /// When using custom environment directions set in [ConferenceService.setSpatialEnvironment], the rotation is defined to always rotate about the relevant axis according to the left handed curl rule. In the animations above you can see, for the y-axis rotation, if you curl your left hand up around with your thumb pointing down the +y axis, the direction the participant will rotate is in the direction the fingers are curling around the given axis. You can see the rotation arrows in those reference animations which correspond to positive rotation direction are pointing the same direction as the fingers of the curled left hand.
 ///
 /// When a direction contains rotations around more than one axis, the rotations are applied in a defined order: yaw, pitch, and then roll. With the standard environment, this corresponds to y, x, and then z. When using custom environment directions, the directions are always in the order of yaw/pitch/roll, but which (x,y,z) axis those correspond to is different.
+///
+/// {@category Models}
 class SpatialDirection {
   /// The Euler rotation about the x-axis, in degrees.
   double x;