+++ lib/src/sdk_api/recording_service.dart
@@ -3,6 +3,8 @@ import '../mapper/mapper.dart';
 import 'models/recording.dart';
 
 /// The RecordingService allows recording conferences.
+///
+/// {@category Services}
 class RecordingService {
   /// @internal
   final _methodChannel =